+++ e2e/e2e-app/src/component-paths.ts
@@ -6,4 +6,5 @@ export const COMPONENT_PATHS = [
   '/special-tags/',
   '/signals/',
   '/disabled-input/',
+  '/component-on-update/',
 ];